Work in progress
Implement low level search filter in tree views

Restore tree state after clearing filter

Python code linting changes detected by black

Make ui reordering icons and view tags in project explorer work again

Add search placeholder help text

Fix format

Disable tree view search feature for QT versions older than 5.10

Store tree state, not filter state to project file

Fix tree state restore
