+++ ApplicationLibCode/ProjectDataModel/WellPath/RimWellPath.cpp
@@ -57,6 +57,7 @@
 #include "RiuMainWindow.h"
 
 #include "cafPdmFieldScriptingCapability.h"
+#include "cafPdmUiTreeAttributes.h"
 #include "cafPdmUiTreeOrdering.h"
 #include "cafPdmUiTreeViewEditor.h"
 #include "cafUtils.h"
