+++ Fwk/AppFwk/cafCommandFeatures/ToggleCommands/cafToggleItemsFeatureImpl.cpp
@@ -194,7 +194,7 @@ caf::PdmUiTreeOrdering* ToggleItemsFeatureImpl::findTreeItemFromSelectedUiItem( 
     if ( pdmUiTreeView )
     {
         QModelIndex modIndex = pdmUiTreeView->findModelIndex( uiItem );
-        return static_cast<caf::PdmUiTreeOrdering*>( modIndex.internalPointer() );
+        return pdmUiTreeView->uiTreeOrderingFromModelIndex( modIndex );
     }
 
     return nullptr;
