+++ Fwk/AppFwk/cafUserInterface/cafPdmUiTreeViewEditor.cpp
@@ -45,6 +45,7 @@
 #include "cafPdmUiDragDropInterface.h"
 #include "cafPdmUiEditorHandle.h"
 #include "cafPdmUiTreeOrdering.h"
+#include "cafPdmUiTreeViewItemDelegate.h"
 #include "cafPdmUiTreeViewQModel.h"
 #include "cafSelectionManager.h"
 
@@ -149,6 +150,7 @@ PdmUiTreeViewEditor::PdmUiTreeViewEditor()
     m_treeView                    = nullptr;
     m_treeViewModel               = nullptr;
     m_delegate                    = nullptr;
+    m_filterModel                 = nullptr;
 }
 
 //--------------------------------------------------------------------------------------------------
@@ -156,6 +158,7 @@ PdmUiTreeViewEditor::PdmUiTreeViewEditor()
 //--------------------------------------------------------------------------------------------------
 PdmUiTreeViewEditor::~PdmUiTreeViewEditor()
 {
+    m_treeView->removeEventFilter( this );
     m_treeViewModel->setPdmItemRoot( nullptr );
 }
 
@@ -170,11 +173,20 @@ QWidget* PdmUiTreeViewEditor::createWidget( QWidget* parent )
     m_mainWidget->setLayout( m_layout );
 
     m_treeViewModel = new caf::PdmUiTreeViewQModel( this );
-    m_treeView      = new PdmUiTreeViewWidget( m_mainWidget );
-    m_treeView->setModel( m_treeViewModel );
+    m_filterModel   = new QSortFilterProxyModel( this );
+    m_filterModel->setFilterKeyColumn( 0 );
+    m_filterModel->setFilterCaseSensitivity( Qt::CaseInsensitive );
+
+#if QT_VERSION >= QT_VERSION_CHECK( 5, 10, 0 )
+    m_filterModel->setRecursiveFilteringEnabled( true );
+#endif
+
+    m_filterModel->setSourceModel( m_treeViewModel );
+    m_treeView = new PdmUiTreeViewWidget( m_mainWidget );
+    m_treeView->setModel( m_filterModel );
     m_treeView->installEventFilter( this );
 
-    m_delegate = new PdmUiTreeViewItemDelegate( this, m_treeViewModel );
+    m_delegate = new PdmUiTreeViewItemDelegate( this, m_filterModel );
 
     m_treeView->setItemDelegate( m_delegate );
 
@@ -256,10 +268,16 @@ void PdmUiTreeViewEditor::selectedUiItems( std::vector<PdmUiItem*>& objects )
     if ( !this->treeView() ) return;
 
     QModelIndexList idxList = this->treeView()->selectionModel()->selectedIndexes();
+    QModelIndexList proxyList;
 
     for ( int i = 0; i < idxList.size(); i++ )
     {
-        caf::PdmUiItem* item = this->m_treeViewModel->uiItemFromModelIndex( idxList[i] );
+        proxyList.append( m_filterModel->mapToSource( idxList[i] ) );
+    }
+
+    for ( int i = 0; i < proxyList.size(); i++ )
+    {
+        caf::PdmUiItem* item = this->m_treeViewModel->uiItemFromModelIndex( proxyList[i] );
         if ( item )
         {
             objects.push_back( item );
@@ -289,8 +307,8 @@ void PdmUiTreeViewEditor::updateMySubTree( PdmUiItem* uiItem )
         }
 
         m_treeViewModel->updateSubTree( itemToUpdate );
-        QModelIndex index = m_treeViewModel->findModelIndex( itemToUpdate );
-        updateItemDelegateForSubTree( index );
+        QModelIndex itemIndex = m_treeViewModel->findModelIndex( itemToUpdate );
+        updateItemDelegateForSubTree( itemIndex );
     }
 }
 
@@ -393,11 +411,11 @@ PdmChildArrayFieldHandle* PdmUiTreeViewEditor::currentChildArrayFieldHandle()
 void PdmUiTreeViewEditor::selectAsCurrentItem( const PdmUiItem* uiItem )
 {
     QModelIndex index        = m_treeViewModel->findModelIndex( uiItem );
-    QModelIndex currentIndex = m_treeView->currentIndex();
+    QModelIndex currentIndex = m_filterModel->mapFromSource( index );
 
     m_treeView->clearSelection();
 
-    m_treeView->setCurrentIndex( index );
+    m_treeView->setCurrentIndex( currentIndex );
 }
 
 //--------------------------------------------------------------------------------------------------
@@ -438,13 +456,34 @@ void PdmUiTreeViewEditor::slotOnSelectionChanged( const QItemSelection& selected
 //--------------------------------------------------------------------------------------------------
 void PdmUiTreeViewEditor::setExpanded( const PdmUiItem* uiItem, bool doExpand ) const
 {
-    QModelIndex index = m_treeViewModel->findModelIndex( uiItem );
-    m_treeView->setExpanded( index, doExpand );
+    QModelIndex index       = m_treeViewModel->findModelIndex( uiItem );
+    QModelIndex filterIndex = m_filterModel->mapFromSource( index );
 
-    if ( doExpand )
+    if ( filterIndex.isValid() )
     {
-        m_treeView->scrollTo( index );
+        m_treeView->setExpanded( filterIndex, doExpand );
+
+        if ( doExpand )
+        {
+            m_treeView->scrollTo( filterIndex );
+        }
     }
+}
+
+//--------------------------------------------------------------------------------------------------
+///
+//--------------------------------------------------------------------------------------------------
+QModelIndex PdmUiTreeViewEditor::mapIndexIfNecessary( QModelIndex index ) const
+{
+    const QAbstractProxyModel* proxyModel = dynamic_cast<const QAbstractProxyModel*>( index.model() );
+
+    QModelIndex returnIndex = index;
+    if ( proxyModel )
+    {
+        returnIndex = proxyModel->mapToSource( index );
+    }
+
+    return returnIndex;
 }
 
 //--------------------------------------------------------------------------------------------------
@@ -452,7 +491,18 @@ void PdmUiTreeViewEditor::setExpanded( const PdmUiItem* uiItem, bool doExpand ) 
 //--------------------------------------------------------------------------------------------------
 PdmUiItem* PdmUiTreeViewEditor::uiItemFromModelIndex( const QModelIndex& index ) const
 {
-    return m_treeViewModel->uiItemFromModelIndex( index );
+    QModelIndex realIndex = mapIndexIfNecessary( index );
+    return m_treeViewModel->uiItemFromModelIndex( realIndex );
+}
+
+//--------------------------------------------------------------------------------------------------
+///
+//--------------------------------------------------------------------------------------------------
+PdmUiTreeOrdering* PdmUiTreeViewEditor::uiTreeOrderingFromModelIndex( const QModelIndex& index ) const
+{
+    QModelIndex realIndex = mapIndexIfNecessary( index );
+    if ( realIndex.isValid() ) return static_cast<caf::PdmUiTreeOrdering*>( realIndex.internalPointer() );
+    return nullptr;
 }
 
 //--------------------------------------------------------------------------------------------------
@@ -460,7 +510,8 @@ PdmUiItem* PdmUiTreeViewEditor::uiItemFromModelIndex( const QModelIndex& index )
 //--------------------------------------------------------------------------------------------------
 QModelIndex PdmUiTreeViewEditor::findModelIndex( const PdmUiItem* object ) const
 {
-    return m_treeViewModel->findModelIndex( object );
+    QModelIndex index = m_treeViewModel->findModelIndex( object );
+    return m_filterModel->mapFromSource( index );
 }
 
 //--------------------------------------------------------------------------------------------------
@@ -469,6 +520,14 @@ QModelIndex PdmUiTreeViewEditor::findModelIndex( const PdmUiItem* object ) const
 void PdmUiTreeViewEditor::setDragDropInterface( PdmUiDragDropInterface* dragDropInterface )
 {
     m_treeViewModel->setDragDropInterface( dragDropInterface );
+}
+
+//--------------------------------------------------------------------------------------------------
+///
+//--------------------------------------------------------------------------------------------------
+void PdmUiTreeViewEditor::setFilterString( QString filterString )
+{
+    m_filterModel->setFilterWildcard( filterString );
 }
 
 //--------------------------------------------------------------------------------------------------
@@ -502,12 +561,15 @@ void PdmUiTreeViewEditor::updateSelectionManager()
 //--------------------------------------------------------------------------------------------------
 ///
 //--------------------------------------------------------------------------------------------------
-void PdmUiTreeViewEditor::updateItemDelegateForSubTree( const QModelIndex& modelIndex /*= QModelIndex()*/ )
+void PdmUiTreeViewEditor::updateItemDelegateForSubTree( const QModelIndex& rootIndex /* = QModelIndex() */ )
 {
-    auto allIndices = m_treeViewModel->allIndicesRecursive();
+    auto allIndices = m_treeViewModel->allIndicesRecursive( rootIndex );
     for ( QModelIndex& index : allIndices )
     {
-        m_delegate->clearTags( index );
+        QModelIndex filterIndex = m_filterModel->mapFromSource( index );
+        if ( !filterIndex.isValid() ) continue;
+
+        m_delegate->clearTags( filterIndex );
 
         PdmUiItem*         uiItem         = m_treeViewModel->uiItemFromModelIndex( index );
         PdmUiObjectHandle* uiObjectHandle = dynamic_cast<PdmUiObjectHandle*>( uiItem );
@@ -523,9 +585,9 @@ void PdmUiTreeViewEditor::updateItemDelegateForSubTree( const QModelIndex& model
                 std::vector<PdmUiItem*> selection;
                 selectedUiItems( selection );
 
-                if ( reorderability && index.row() >= 0 && selection.size() == 1u && selection.front() == uiItem )
+                if ( reorderability && filterIndex.row() >= 0 && selection.size() == 1u && selection.front() == uiItem )
                 {
-                    size_t indexInParent = static_cast<size_t>( index.row() );
+                    size_t indexInParent = static_cast<size_t>( filterIndex.row() );
                     {
                         auto tag          = PdmUiTreeViewItemAttribute::Tag::create();
                         tag->icon         = caf::IconProvider( ":/caf/Up16x16.png" );
@@ -539,7 +601,7 @@ void PdmUiTreeViewEditor::updateItemDelegateForSubTree( const QModelIndex& model
                             tag->icon.setActive( false );
                         }
 
-                        m_delegate->addTag( index, std::move( tag ) );
+                        m_delegate->addTag( filterIndex, std::move( tag ) );
                     }
                     {
                         auto tag          = PdmUiTreeViewItemAttribute::Tag::create();
@@ -553,7 +615,7 @@ void PdmUiTreeViewEditor::updateItemDelegateForSubTree( const QModelIndex& model
                         {
                             tag->icon.setActive( false );
                         }
-                        m_delegate->addTag( index, std::move( tag ) );
+                        m_delegate->addTag( filterIndex, std::move( tag ) );
                     }
                 }
             }
@@ -564,7 +626,7 @@ void PdmUiTreeViewEditor::updateItemDelegateForSubTree( const QModelIndex& model
             {
                 if ( !tag->text.isEmpty() || tag->icon.valid() )
                 {
-                    m_delegate->addTag( index, std::move( tag ) );
+                    m_delegate->addTag( filterIndex, std::move( tag ) );
                 }
             }
         }
@@ -587,260 +649,4 @@ bool PdmUiTreeViewEditor::isAppendOfClassNameToUiItemTextEnabled()
     return m_appendClassNameToUiItemText;
 }
 
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-PdmUiTreeViewItemDelegate::PdmUiTreeViewItemDelegate( PdmUiTreeViewEditor* parent, PdmUiTreeViewQModel* model )
-    : QStyledItemDelegate( parent->treeView() )
-    , m_treeView( parent )
-    , m_model( model )
-{
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-void PdmUiTreeViewItemDelegate::clearTags( QModelIndex index )
-{
-    m_tags.erase( index );
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-void PdmUiTreeViewItemDelegate::clearAllTags()
-{
-    m_tags.clear();
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-void PdmUiTreeViewItemDelegate::addTag( QModelIndex index, std::unique_ptr<PdmUiTreeViewItemAttribute::Tag> tag )
-{
-    std::vector<std::unique_ptr<PdmUiTreeViewItemAttribute::Tag>>& tagList = m_tags[index];
-    tagList.push_back( std::move( tag ) );
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-QRect PdmUiTreeViewItemDelegate::tagRect( const QRect& itemRect, QModelIndex index, size_t tagIndex ) const
-{
-    auto it = m_tags.find( index );
-    if ( it == m_tags.end() ) return QRect();
-
-    QSize fullSize = itemRect.size();
-
-    QPoint offset( 0, 0 );
-
-    for ( size_t i = 0; i < it->second.size(); ++i )
-    {
-        const PdmUiTreeViewItemAttribute::Tag* tag = it->second[i].get();
-        if ( tag->icon.valid() )
-        {
-            auto  icon     = tag->icon.icon();
-            QSize iconSize = icon->actualSize( fullSize );
-            QRect iconRect;
-            if ( tag->position == PdmUiTreeViewItemAttribute::Tag::AT_END )
-            {
-                QPoint bottomRight = itemRect.bottomRight() - offset;
-                QPoint topLeft     = bottomRight - QPoint( iconSize.width(), iconSize.height() );
-                iconRect           = QRect( topLeft, bottomRight );
-            }
-            else
-            {
-                QPoint topLeft     = itemRect.topLeft() + offset;
-                QPoint bottomRight = topLeft + QPoint( iconSize.width(), iconSize.height() );
-                iconRect           = QRect( topLeft, bottomRight );
-            }
-            offset += QPoint( iconSize.width() + 2, 0 );
-
-            if ( i == tagIndex ) return iconRect;
-        }
-    }
-    return QRect();
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-void PdmUiTreeViewItemDelegate::paint( QPainter* painter, const QStyleOptionViewItem& option, const QModelIndex& index ) const
-{
-    QStyledItemDelegate::paint( painter, option, index );
-
-    auto it = m_tags.find( index );
-    if ( it == m_tags.end() ) return;
-
-    // Save painter so we can restore it
-    painter->save();
-
-    QRect  rect   = option.rect;
-    QPoint center = rect.center();
-
-    QSize fullSize = rect.size();
-
-    QPoint offset( 0, 0 );
-
-    for ( const std::unique_ptr<PdmUiTreeViewItemAttribute::Tag>& tag : it->second )
-    {
-        if ( tag->selectedOnly && !( option.state & QStyle::State_Selected ) ) continue;
-
-        if ( tag->icon.valid() )
-        {
-            auto  icon     = tag->icon.icon();
-            QSize iconSize = icon->actualSize( fullSize );
-            QRect iconRect;
-            if ( tag->position == PdmUiTreeViewItemAttribute::Tag::AT_END )
-            {
-                QPoint bottomRight( rect.bottomRight().x() - offset.x(), center.y() + iconSize.height() / 2 );
-                QPoint topLeft( bottomRight.x() - iconSize.width(), bottomRight.y() - iconSize.height() );
-                iconRect = QRect( topLeft, bottomRight );
-            }
-            else
-            {
-                QPoint topLeft( rect.topLeft().x() + offset.x(), center.y() - iconSize.height() / 2 );
-                QPoint bottomRight( topLeft.x() + iconSize.width(), topLeft.y() + iconSize.height() );
-                iconRect = QRect( topLeft, bottomRight );
-            }
-            offset += QPoint( iconSize.width() + 2, 0 );
-            icon->paint( painter, iconRect );
-        }
-        else
-        {
-            const int insideTopBottomMargins  = 1;
-            const int insideleftRightMargins  = 6;
-            const int outsideLeftRightMargins = 4;
-
-            QFont font = QApplication::font();
-            if ( font.pixelSize() > 0 )
-            {
-                font.setPixelSize( std::max( 1, font.pixelSize() - 1 ) );
-            }
-            else
-            {
-                font.setPointSize( std::max( 1, font.pointSize() - 1 ) );
-            }
-            painter->setFont( font );
-
-            QString text    = tag->text;
-            QColor  bgColor = tag->bgColor;
-            QColor  fgColor = tag->fgColor;
-
-            QSize textSize( QFontMetrics( font ).size( Qt::TextSingleLine, text ) );
-            int   textDiff = ( fullSize.height() - textSize.height() );
-
-            QRect textRect;
-            if ( tag->position == PdmUiTreeViewItemAttribute::Tag::AT_END )
-            {
-                QPoint bottomRight     = rect.bottomRight() - QPoint( outsideLeftRightMargins, 0 ) - offset;
-                QPoint textBottomRight = bottomRight - QPoint( insideleftRightMargins, textDiff / 2 );
-                QPoint textTopLeft     = textBottomRight - QPoint( textSize.width(), textSize.height() );
-                textRect               = QRect( textTopLeft, textBottomRight );
-            }
-            else
-            {
-                QPoint textTopLeft = QPoint( 0, rect.topLeft().y() ) + offset +
-                                     QPoint( outsideLeftRightMargins + insideleftRightMargins, +textDiff / 2 );
-                QPoint textBottomRight = textTopLeft + QPoint( textSize.width(), textSize.height() );
-                textRect               = QRect( textTopLeft, textBottomRight );
-            }
-
-            QRect tagRect = textRect.marginsAdded(
-                QMargins( insideleftRightMargins, insideTopBottomMargins, insideleftRightMargins, insideTopBottomMargins ) );
-
-            offset += QPoint( tagRect.width() + 2, 0 );
-
-            QBrush brush( bgColor );
-
-            painter->setBrush( brush );
-            painter->setPen( bgColor );
-            painter->setRenderHint( QPainter::Antialiasing );
-            const double xRoundingRadiusPercent = 50.0;
-            const double yRoundingRadiusPercent = 25.0;
-            painter->drawRoundedRect( tagRect, xRoundingRadiusPercent, yRoundingRadiusPercent, Qt::RelativeSize );
-
-            painter->setPen( fgColor );
-            painter->drawText( textRect, Qt::AlignCenter, text );
-        }
-    }
-    // Restore painter
-    painter->restore();
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-std::vector<const caf::PdmUiTreeViewItemAttribute::Tag*> PdmUiTreeViewItemDelegate::tags( QModelIndex index ) const
-{
-    std::vector<const caf::PdmUiTreeViewItemAttribute::Tag*> tagPtrVector;
-
-    auto it = m_tags.find( index );
-    if ( it != m_tags.end() )
-    {
-        for ( const auto& tag : it->second )
-        {
-            tagPtrVector.push_back( tag.get() );
-        }
-    }
-    return tagPtrVector;
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-bool PdmUiTreeViewItemDelegate::editorEvent( QEvent*                     event,
-                                             QAbstractItemModel*         model,
-                                             const QStyleOptionViewItem& option,
-                                             const QModelIndex&          itemIndex )
-{
-    if ( event->type() == QEvent::MouseButtonPress )
-    {
-        QMouseEvent* mouseEvent = static_cast<QMouseEvent*>( event );
-
-        if ( mouseEvent->button() == Qt::LeftButton && mouseEvent->modifiers() == Qt::NoModifier )
-        {
-            const PdmUiTreeViewItemAttribute::Tag* tag;
-            if ( tagClicked( mouseEvent->pos(), option.rect, itemIndex, &tag ) )
-            {
-                QModelIndex parentIndex = itemIndex.parent();
-
-                auto uiItem       = m_treeView->uiItemFromModelIndex( itemIndex );
-                auto parentUiItem = m_treeView->uiItemFromModelIndex( parentIndex );
-
-                tag->clicked.send( (size_t)itemIndex.row() );
-
-                m_treeView->updateSubTree( parentUiItem );
-                m_treeView->selectAsCurrentItem( uiItem );
-            }
-        }
-    }
-
-    return QStyledItemDelegate::editorEvent( event, model, option, itemIndex );
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-bool PdmUiTreeViewItemDelegate::tagClicked( const QPoint&                           pos,
-                                            const QRect&                            itemRect,
-                                            const QModelIndex&                      itemIndex,
-                                            const PdmUiTreeViewItemAttribute::Tag** tag ) const
-{
-    if ( itemIndex.isValid() )
-    {
-        auto itemTags = tags( itemIndex );
-
-        for ( size_t i = 0; i < itemTags.size(); ++i )
-        {
-            QRect rect = tagRect( itemRect, itemIndex, i );
-            if ( rect.contains( pos ) )
-            {
-                *tag = itemTags[i];
-                return true;
-            }
-        }
-    }
-    return false;
-}
 } // end namespace caf
